backend/requirements.txt
Django==1.9.1
PyYAML==3.11
SQLAlchemy==1.0.4
argparse==1.4.0
coverage==4.0.3
dj-database-url==0.3.0
django-bootstrap3==7.0.1
django-fontawesome==0.2.6
django-jquery==1.12.2
django-nose==1.4.3
django-pipeline==1.6.8
django-postgrespool==0.3.0
extras==0.0.3
fixtures==2.0.0
funcsigs==1.0.2
fuzzywuzzy==0.10.0
lettuce==0.2.21
linecache2==1.0.0
mock==2.0.0
nose==1.3.7
pbr==1.9.1
psycopg2==2.6
python-mimeparse==1.5.2
python-subunit==1.2.0
six
sure==1.2.24
testtools==2.1.0
traceback2==1.4.0
unittest2==1.1.0
virtualenv
wsgiref==0.1.2
factory_boy==2.7.0
djangorestframework
django-cors-headers